+++ spec/rack/twilio_webhook_authentication_spec.rb
@@ -18,6 +18,28 @@ describe Rack::TwilioWebhookAuthentication do
         Rack::TwilioWebhookAuthentication.new(@app, 'ABC', /\/voice/, /\/sms/)
       }.not_to raise_error
     end
+
+    it 'should initialize with an app, dynamic token and paths' do
+      expect {
+        Rack::TwilioWebhookAuthentication.new(@app, nil, /\/voice/, /\/sms/)
+      }.not_to raise_error
+    end
+  end
+
+  describe 'calling against one path with dynamic auth token' do
+    it 'should allow a request through if it validates' do
+      auth_token = 'qwerty'
+      account_sid = 12345
+      expect_any_instance_of(Rack::Request).to receive(:post?).and_return(true)
+      expect_any_instance_of(Rack::Request).to receive(:POST).and_return({'AccountSid' => account_sid})
+      @middleware = Rack::TwilioWebhookAuthentication.new(@app, nil, /\/voice/) { |asid| auth_token}
+      request_validator = double('RequestValidator')
+      expect(Twilio::Util::RequestValidator).to receive(:new).with(auth_token).and_return(request_validator)
+      expect(request_validator).to receive(:validate).and_return(true)
+      request = Rack::MockRequest.env_for('/voice')
+      status, headers, body = @middleware.call(request)
+      expect(status).to be(200)
+    end
   end
 
   describe 'calling against one path' do