+++ spec/rest/account_spec.rb
@@ -78,4 +78,12 @@ describe Twilio::REST::Account do
       'someUri/Tokens'
     )
   end
+
+  it 'sets up an addresses resources object' do
+    expect(@account).to respond_to(:addresses)
+    expect(@account.addresses.instance_variable_get('@path')).to eq(
+      'someUri/Addresses'
+    )
+  end
+
 end